Revert "投稿機能を実装"
