+++ bbsApp/modules/authenticator.js
@@ -5,7 +5,7 @@ const jwtConfig = {
   secret: 'secret_key',
   options: {
     algorithm: 'HS256',
-    expiresIn: '10m'
+    expiresIn: '10s'
   }
 }
 
@@ -13,7 +13,6 @@ module.exports = {
   // tokenを生成してリクエストヘッダーに追加
   getToken(req, res, next) {
     const payload = {
-      id: req.body.userId,
       name: req.body.name,
       email: req.body.email
     }
@@ -31,7 +30,6 @@ module.exports = {
       if (error) throw new Error('invalid token');
       if (user.name === req.body.name) {
         res.user = {
-          userId: user.id,
           name: user.name,
           token: token
         };
@@ -46,15 +44,15 @@ module.exports = {
   checkToken(req, res, next) {
     if (!req.session.token) {
       res.redirect('/auth/login');
+      return
     } else {
       jwt.verify(req.session.token, jwtConfig.secret, (error, user) => {
         if (error) {
           res.redirect('/auth/login');
-          return ;
+          return
         }
         if (user.name === req.session.username) {
           res.user = {
-            userId: user.id,
             name: user.name,
             token: req.session.token
           };
