--- bbsApp/routes/likeRoute.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const likeController = require('../controllers/likeController');
-
-module.exports = [
-  router.post('/likes', likeController.checkLike),
-]
