+++ bbsApp/views/components/header.ejs
@@ -3,6 +3,5 @@
   <head>
     <title><%= title %></title>
     <link rel="stylesheet" href="/stylesheets/bootstrap.min.css">
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.0/css/all.min.css" rel="stylesheet">
   </head>
   <body>
