+++ bbsApp/views/index.ejs
@@ -1,0 +1,3 @@
+<%- include('components/header'); %>
+<%- include('components/navbar'); %>
+<%- include('components/footer'); %>
