+++ bbsApp/views/user/allPosts.ejs
@@ -1,0 +1,4 @@
+<%- include('../components/header'); %>
+<%- include('../components/navbar_logged_in'); %>
+<h1>みんなの投稿</h1>
+<%- include('../components/footer'); %>
