+++ bbsApp/views/user/myPost.ejs
@@ -1,0 +1,4 @@
+<%- include('../components/header'); %>
+<%- include('../components/navbar_logged_in'); %>
+<h1>投稿を作成</h1>
+<%- include('../components/footer'); %>
